requirements.txt
# Core dependencies
annotated-types==0.7.0
cachetools==5.5.2
certifi==2025.1.31
click==8.1.8
google-auth==2.38.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dotenv==1.1.0
rsa==4.9
setuptools==76.0.0
sniffio==1.3.1
urllib3==2.3.0
uv==0.6.6
wheel==0.45.1

# Pydantic - compatible with ag-ui-protocol (>=2.11.2) and genai-prices (>=2.11.7)
pydantic>=2.11.7,<3.0.0
pydantic-settings==2.8.1
pydantic_core>=2.31.1

# Typing - compatible with cryptography (>=4.13.2 for Python < 3.11)
typing-inspection==0.4.0
typing_extensions>=4.13.2

# Google ADK and dependencies - compatible with google-adk 1.14.1
google-genai>=1.21.1,<2.0.0
mcp>=1.8.0,<2.0.0

# Web framework - compatible with google-adk 1.14.1
starlette>=0.46.2,<1.0.0
sse-starlette==2.2.1
uvicorn==0.34.0

# Requests - compatible with google-adk 1.14.1
requests>=2.32.4,<3.0.0

# WebSockets - compatible with google-adk 1.14.1
websockets>=15.0.1,<16.0.0

# Protobuf - compatible with temporalio (<6)
protobuf>=3.20,<6

# Database - async support
asyncpg==0.29.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

google-adk==1.14.1
aiohttp